requirements.txt
# CORE DE PROYECTO BACKEND Y FRAMEWORKS
fastapi==0.112.0
fastapi-cli==0.0.4
fastapi-sessions==0.3.2
starlette>=0.37.2,<0.38.0
uvicorn==0.22.0
gunicorn==21.2.0
Werkzeug==2.1.2
jinja2==3.1.2
itsdangerous==2.1.2
python-multipart==0.0.9
aiofiles==23.1.0

# DB
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32  # opcional si quieres ORM

# Utilidades
python-dotenv==1.0.1

# QR
qrcode==7.4.2
Pillow==10.4.0

# Azure (si vas a subir los PNG)
azure-storage-blob==12.19.1

# Opcional
tqdm==4.66.5
openpyxl==3.1.5
